# api/app.py
"""The core API: the fleet registry and its audit log. P2-05, P2-06.

Orders and pricing (P2-03, P2-04, P2-07) are not here: they are the courier
business, and that decision is open (2026-09-28).

**Every route needs a signed-in operator** (P6-08, `api/auth_http.py`):
reading needs `viewer`, changing the registry needs `admin`. Every change is
recorded in `events` against the operator who made it. The API still binds
to loopback by default; P0-09 exposes it through TLS.

The OpenAPI schema is generated by FastAPI at `/openapi.json`; TypeScript
types are generated from it (CLAUDE.md), never written by hand.
"""

# No `from __future__ import annotations` here, deliberately: route
# signatures use `Annotated[Operator, Depends(viewer)]` where `viewer` is a
# local of the factory, and a postponed annotation is evaluated against
# module globals, where it does not exist. FastAPI then silently treats the
# parameter as a required query string called `_`.

from datetime import UTC, datetime, timedelta
from pathlib import Path
from typing import Annotated, Any
from uuid import UUID

from fastapi import Depends, FastAPI, HTTPException, Query
from fastapi.responses import HTMLResponse, RedirectResponse
from fastapi.staticfiles import StaticFiles
from pydantic import BaseModel, Field

from api.assets import STATIC, mount_map_assets
from api.auth import Operator, Role
from api.auth_http import AccountStore, Authenticator, auth_router, require
from api.registry import (
    AirframeParams,
    ConflictError,
    DroneStatus,
    FleetRegistry,
    NotFoundError,
    PilotStatus,
    RegistryError,
)
from api.replay import DroneNotFoundError, ReplayError, ReplayStore, WindowTooLargeError
from api.zones import ZoneReader

MAX_EVENTS_PER_PAGE = 1_000
MAX_FLIGHTS_PER_PAGE = 200
# How far back the flight list looks when not told. A default, not a limit.
DEFAULT_FLIGHT_LOOKBACK = timedelta(days=30)


class BaseIn(BaseModel):
    name: str = Field(min_length=1)
    lat_deg: float = Field(ge=-90, le=90)
    lon_deg: float = Field(ge=-180, le=180)
    elevation_amsl_m: float | None = None
    capacity: int = Field(default=0, ge=0)
    charging_slots: int = Field(default=0, ge=0)


class BaseOut(BaseModel):
    id: UUID
    name: str
    lat_deg: float
    lon_deg: float
    elevation_amsl_m: float | None
    capacity: int
    charging_slots: int
    created_at: datetime


class PilotIn(BaseModel):
    name: str = Field(min_length=1)
    license_ref: str | None = None
    max_concurrent_drones: int = Field(default=1, ge=1)


class PilotOut(BaseModel):
    id: UUID
    name: str
    license_ref: str | None
    status: PilotStatus
    max_concurrent_drones: int
    created_at: datetime


class PilotStatusIn(BaseModel):
    status: PilotStatus


class DroneIn(BaseModel):
    serial: str = Field(min_length=1)
    label: str = Field(min_length=1)
    model: str | None = None
    max_payload_g: int | None = Field(default=None, gt=0)
    max_range_m: float | None = Field(default=None, gt=0)
    battery_capacity_wh: float | None = Field(default=None, gt=0)
    cruise_speed_ms: float | None = Field(default=None, gt=0)
    avg_power_w: float | None = Field(default=None, gt=0)
    home_base_id: UUID | None = None
    current_pilot_id: UUID | None = None


class DroneOut(BaseModel):
    id: UUID
    serial: str
    label: str
    model: str | None
    max_payload_g: int | None
    max_range_m: float | None
    battery_capacity_wh: float | None
    cruise_speed_ms: float | None
    avg_power_w: float | None
    home_base_id: UUID | None
    current_pilot_id: UUID | None
    in_maintenance: bool
    retired_at: datetime | None
    created_at: datetime
    # Derived from telemetry on every read; see `api.registry.derive_status`.
    status: DroneStatus


class MaintenanceIn(BaseModel):
    in_maintenance: bool


class ZoneOut(BaseModel):
    id: UUID
    name: str
    # no_fly, restricted, corridor or base (migration 0001_fleet).
    type: str
    min_alt_amsl_m: float | None
    max_alt_amsl_m: float | None
    # A GeoJSON Polygon in WGS84, as PostGIS writes it.
    geometry: dict[str, Any]


class EventOut(BaseModel):
    id: int
    ts: datetime
    actor_type: str
    actor_id: str | None
    entity_type: str
    entity_id: str
    event_type: str
    payload: dict[str, Any]


def _http(error: RegistryError) -> HTTPException:
    if isinstance(error, NotFoundError):
        return HTTPException(status_code=404, detail=str(error))
    if isinstance(error, ConflictError):
        return HTTPException(status_code=409, detail=str(error))
    return HTTPException(status_code=400, detail=str(error))


def _replay_http(error: ReplayError) -> HTTPException:
    if isinstance(error, DroneNotFoundError):
        return HTTPException(status_code=404, detail=str(error))
    if isinstance(error, WindowTooLargeError):
        return HTTPException(status_code=413, detail=str(error))
    return HTTPException(status_code=422, detail=str(error))


def _utc(value: datetime) -> datetime:
    """A time with no zone is refused rather than guessed (CLAUDE.md: UTC)."""
    if value.tzinfo is None:
        raise HTTPException(
            status_code=422, detail="times must carry a zone, e.g. a trailing Z"
        )
    return value.astimezone(UTC)


def create_api_app(
    registry: FleetRegistry,
    *,
    auth: AccountStore,
    feed_secret: bytes,
    feed_ticket_ttl_s: float,
    cookie_secure: bool = True,
    replay: ReplayStore | None = None,
    basemap_dir: Path | None = None,
    console_feed_url: str | None = None,
    console_app_dir: Path | None = None,
) -> FastAPI:
    """The API. `auth` is required: there is no way to build it open.

    It also serves the operator pages - sign-in, the live map, replay - with
    the map libraries and base map they need. The pages carry no data; every
    request they make for data is authenticated.
    """
    app = FastAPI(title="courier API", version="0.1.0")
    viewer = require(auth, Role.VIEWER)
    admin = require(auth, Role.ADMIN)
    app.include_router(
        auth_router(
            auth,
            feed_secret=feed_secret,
            feed_ticket_ttl_s=feed_ticket_ttl_s,
            cookie_secure=cookie_secure,
        )
    )

    # --- bases ----------------------------------------------------------------

    @app.post("/bases", response_model=BaseOut, status_code=201)
    async def create_base(
        body: BaseIn, operator: Annotated[Operator, Depends(admin)]
    ) -> dict[str, Any]:
        try:
            return await registry.create_base(**body.model_dump(), actor=operator.actor)
        except RegistryError as error:
            raise _http(error) from error

    @app.get("/bases", response_model=list[BaseOut])
    async def list_bases(
        _: Annotated[Operator, Depends(viewer)],
    ) -> list[dict[str, Any]]:
        return await registry.list_bases()

    # --- pilots ---------------------------------------------------------------

    @app.post("/pilots", response_model=PilotOut, status_code=201)
    async def create_pilot(
        body: PilotIn, operator: Annotated[Operator, Depends(admin)]
    ) -> dict[str, Any]:
        try:
            return await registry.create_pilot(
                **body.model_dump(), actor=operator.actor
            )
        except RegistryError as error:
            raise _http(error) from error

    @app.get("/pilots", response_model=list[PilotOut])
    async def list_pilots(
        _: Annotated[Operator, Depends(viewer)],
    ) -> list[dict[str, Any]]:
        return await registry.list_pilots()

    @app.put("/pilots/{pilot_id}/status", response_model=PilotOut)
    async def set_pilot_status(
        pilot_id: UUID,
        body: PilotStatusIn,
        operator: Annotated[Operator, Depends(admin)],
    ) -> dict[str, Any]:
        try:
            return await registry.set_pilot_status(
                pilot_id, body.status, actor=operator.actor
            )
        except RegistryError as error:
            raise _http(error) from error

    # --- drones ---------------------------------------------------------------

    @app.post("/drones", response_model=DroneOut, status_code=201)
    async def register_drone(
        body: DroneIn, operator: Annotated[Operator, Depends(admin)]
    ) -> dict[str, Any]:
        fields = body.model_dump()
        params = AirframeParams(
            max_payload_g=fields.pop("max_payload_g"),
            max_range_m=fields.pop("max_range_m"),
            battery_capacity_wh=fields.pop("battery_capacity_wh"),
            cruise_speed_ms=fields.pop("cruise_speed_ms"),
            avg_power_w=fields.pop("avg_power_w"),
        )
        try:
            return await registry.register_drone(
                params=params, **fields, actor=operator.actor
            )
        except RegistryError as error:
            raise _http(error) from error

    @app.get("/drones", response_model=list[DroneOut])
    async def list_drones(
        _: Annotated[Operator, Depends(viewer)], include_retired: bool = False
    ) -> list[dict[str, Any]]:
        return await registry.list_drones(include_retired=include_retired)

    @app.get("/drones/{drone_id}", response_model=DroneOut)
    async def get_drone(
        drone_id: UUID, _: Annotated[Operator, Depends(viewer)]
    ) -> dict[str, Any]:
        try:
            return await registry.get_drone(drone_id)
        except RegistryError as error:
            raise _http(error) from error

    @app.put("/drones/{drone_id}/maintenance", response_model=DroneOut)
    async def set_maintenance(
        drone_id: UUID,
        body: MaintenanceIn,
        operator: Annotated[Operator, Depends(admin)],
    ) -> dict[str, Any]:
        try:
            return await registry.set_maintenance(
                drone_id, body.in_maintenance, actor=operator.actor
            )
        except RegistryError as error:
            raise _http(error) from error

    @app.post("/drones/{drone_id}/retire", response_model=DroneOut)
    async def retire_drone(
        drone_id: UUID, operator: Annotated[Operator, Depends(admin)]
    ) -> dict[str, Any]:
        try:
            return await registry.retire_drone(drone_id, actor=operator.actor)
        except RegistryError as error:
            raise _http(error) from error

    # --- airspace (P6-01) ------------------------------------------------------

    zone_reader = ZoneReader(engine=registry.engine) if registry is not None else None

    @app.get("/airspace/zones", response_model=list[ZoneOut])
    async def zones(
        _: Annotated[Operator, Depends(viewer)],
    ) -> list[dict[str, Any]]:
        """Every zone, for drawing. The monitor alerts on no-fly and restricted."""
        if zone_reader is None:
            raise HTTPException(status_code=503, detail="no relational database")
        return await zone_reader.zones()

    # --- audit (P2-06) --------------------------------------------------------

    @app.get("/events", response_model=list[EventOut])
    async def events(
        _: Annotated[Operator, Depends(viewer)],
        entity_type: str | None = None,
        entity_id: str | None = None,
        since: datetime | None = None,
        until: datetime | None = None,
        after_id: int | None = None,
        limit: int = Query(default=500, ge=1, le=MAX_EVENTS_PER_PAGE),
    ) -> list[dict[str, Any]]:
        """Filter by entity and time; page with `after_id` = the last `id` seen."""
        return await registry.events(
            entity_type=entity_type,
            entity_id=entity_id,
            since=since,
            until=until,
            after_id=after_id,
            limit=limit,
        )

    # --- operator pages --------------------------------------------------------

    mount_map_assets(app, basemap_dir)

    @app.get("/map", response_class=HTMLResponse, include_in_schema=False)
    async def map_page() -> str:
        """The live map. Its feed comes from the console, with a ticket from
        here (P6-08)."""
        return (STATIC / "map.html").read_text(encoding="utf-8")

    # P6-01. The operator console, built from web-pilot/. Absent until built;
    # the minimal /map stays as the fallback.
    has_console = (
        console_app_dir is not None and (console_app_dir / "index.html").is_file()
    )
    if has_console:
        app.mount(
            "/app", StaticFiles(directory=console_app_dir, html=True), name="console"
        )

    @app.get("/", include_in_schema=False)
    async def home() -> RedirectResponse:
        return RedirectResponse("/app/" if has_console else "/map")

    @app.get("/config.json", include_in_schema=False)
    async def page_config() -> dict[str, Any]:
        """Where the pages find the console feed. Not secret."""
        return {"console_feed_url": console_feed_url}

    # --- replay (P10-03) -------------------------------------------------------

    if replay is not None:
        _add_replay_routes(app, replay, auth)

    return app


def _add_replay_routes(app: FastAPI, replay: ReplayStore, auth: Authenticator) -> None:
    viewer = require(auth, Role.VIEWER)

    @app.get("/replay", response_class=HTMLResponse, include_in_schema=False)
    async def replay_page() -> str:
        return (STATIC / "replay.html").read_text(encoding="utf-8")

    @app.get("/replay/drones")
    async def replay_drones(
        _: Annotated[Operator, Depends(viewer)],
    ) -> list[dict[str, Any]]:
        """Every aircraft with telemetry that can be replayed, retired or not."""
        return await replay.drones()

    @app.get("/replay/drones/{drone_id}/flights")
    async def flights(
        drone_id: UUID,
        _: Annotated[Operator, Depends(viewer)],
        since: datetime | None = None,
        until: datetime | None = None,
        limit: int = Query(default=50, ge=1, le=MAX_FLIGHTS_PER_PAGE),
    ) -> list[dict[str, Any]]:
        """Armed spans of this aircraft, newest first."""
        end = _utc(until) if until is not None else datetime.now(tz=UTC)
        start = _utc(since) if since is not None else end - DEFAULT_FLIGHT_LOOKBACK
        try:
            return await replay.flights(drone_id, since=start, until=end, limit=limit)
        except ReplayError as error:
            raise _replay_http(error) from error

    @app.get("/replay/drones/{drone_id}")
    async def replay_window(
        drone_id: UUID,
        start: datetime,
        end: datetime,
        _: Annotated[Operator, Depends(viewer)],
    ) -> dict[str, Any]:
        """Every stored sample in the window, the holes in it and why, the
        loss evidence around it, and the airspace alerts raised and cleared."""
        try:
            return await replay.replay(drone_id, start=_utc(start), end=_utc(end))
        except ReplayError as error:
            raise _replay_http(error) from error