# web-pilot/openapi.json
{
  "components": {
    "schemas": {
      "BaseIn": {
        "properties": {
          "capacity": {
            "default": 0,
            "minimum": 0.0,
            "title": "Capacity",
            "type": "integer"
          },
          "charging_slots": {
            "default": 0,
            "minimum": 0.0,
            "title": "Charging Slots",
            "type": "integer"
          },
          "elevation_amsl_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation Amsl M"
          },
          "lat_deg": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat Deg",
            "type": "number"
          },
          "lon_deg": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon Deg",
            "type": "number"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat_deg",
          "lon_deg"
        ],
        "title": "BaseIn",
        "type": "object"
      },
      "BaseOut": {
        "properties": {
          "capacity": {
            "title": "Capacity",
            "type": "integer"
          },
          "charging_slots": {
            "title": "Charging Slots",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elevation_amsl_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation Amsl M"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lat_deg": {
            "title": "Lat Deg",
            "type": "number"
          },
          "lon_deg": {
            "title": "Lon Deg",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "lat_deg",
          "lon_deg",
          "elevation_amsl_m",
          "capacity",
          "charging_slots",
          "created_at"
        ],
        "title": "BaseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MaintenanceIn": {
        "properties": {
          "in_maintenance": {
            "title": "In Maintenance",
            "type": "boolean"
          }
        },
        "required": [
          "in_maintenance"
        ],
        "title": "MaintenanceIn",
        "type": "object"
      },
...
      "PilotStatus": {
        "enum": [
          "AVAILABLE",
          "ON_DUTY",
          "OFF_DUTY"
        ],
        "title": "PilotStatus",
        "type": "string"
      },
      "PilotStatusIn": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/PilotStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "PilotStatusIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "courier API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bases": {
      "get": {
        "operationId": "list_bases_bases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BaseOut"
                  },
                  "title": "Response List Bases Bases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Bases"
      },
      "post": {
        "operationId": "create_base_bases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BaseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Base"
      }
    },
    "/drones": {
      "get": {
        "operationId": "list_drones_drones_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_retired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retired",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DroneOut"
                  },
                  "title": "Response List Drones Drones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Drones"
      },
      "post": {
        "operationId": "register_drone_drones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DroneIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Drone"
      }
    },
    "/drones/{drone_id}": {
      "get": {
        "operationId": "get_drone_drones__drone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Drone"
      }
    },
    "/drones/{drone_id}/maintenance": {
      "put": {
        "operationId": "set_maintenance_drones__drone_id__maintenance_put",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Maintenance"
      }
    },
    "/drones/{drone_id}/retire": {
      "post": {
        "operationId": "retire_drone_drones__drone_id__retire_post",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DroneOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire Drone"
      }
    },
    "/events": {
      "get": {
        "description": "Filter by entity and time; page with `after_id` = the last `id` seen.",
        "operationId": "events_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "entity_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Events Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events"
      }
    },
...
    "/pilots": {
      "get": {
        "operationId": "list_pilots_pilots_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PilotOut"
                  },
                  "title": "Response List Pilots Pilots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Pilots"
      },
      "post": {
        "operationId": "create_pilot_pilots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pilot"
      }
    },
    "/pilots/{pilot_id}/status": {
      "put": {
        "operationId": "set_pilot_status_pilots__pilot_id__status_put",
        "parameters": [
          {
            "in": "path",
            "name": "pilot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Pilot Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PilotStatusIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Pilot Status"
      }
    },
...
    "/replay/drones/{drone_id}": {
      "get": {
        "description": "Every stored sample in the window, the holes in it and why, the\nloss evidence around it, and the airspace alerts raised and cleared.",
        "operationId": "replay_window_replay_drones__drone_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Replay Window Replay Drones  Drone Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Window"
      }
    },
    "/replay/drones/{drone_id}/flights": {
      "get": {
        "description": "Armed spans of this aircraft, newest first.",
        "operationId": "flights_replay_drones__drone_id__flights_get",
        "parameters": [
          {
            "in": "path",
            "name": "drone_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Drone Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Flights Replay Drones  Drone Id  Flights Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Flights"
      }
    }
  }
}